hackathon-v1/vinfast-service-copilot-demo/requirements.txt
langgraph
langchain
langchain-core
langchain-ollama
langchain-google-genai
streamlit
python-dotenv
pydantic
requests